Move to `nn/layer/loss.py`
test=develop
