+++ python/paddle/nn/layer/loss.py
@@ -17,7 +17,7 @@ import paddle.fluid as fluid
 __all__ = [
     #'NCELoss',
     'CrossEntropyLoss',
-    #    'MSELoss',
+    'MSELoss',
     'L1Loss',
     #    'NLLLoss',
     'BCELoss'
@@ -133,6 +133,122 @@ class CrossEntropyLoss(fluid.dygraph.Layer):
             return fluid.layers.reduce_mean(out)
         else:
             return out
+
+
+class MSELoss(fluid.dygraph.layers.Layer):
+    """
+    **Mean Square Error Loss**
+    Computes the mean square error (squared L2 norm) of given input and label.
+
+    If :attr:`reduction` is set to ``'none'``, loss is calculated as:
+
+    .. math::
+        Out = (input - label)^2
+
+    If :attr:`reduction` is set to ``'mean'``, loss is calculated as:
+
+    .. math::
+        Out = \operatorname{mean}((input - label)^2)
+
+    If :attr:`reduction` is set to ``'sum'``, loss is calculated as:
+
+    .. math::
+        Out = \operatorname{sum}((input - label)^2)
+
+    where `input` and `label` are `float32` tensors of arbitrary shapes.
+
+    Parameters:
+        reduction (string, optional): The reduction method for the output,
+            could be 'none' | 'mean' | 'sum'.
+            'none': no reduction will be applied
+            'mean': the output will be averaged
+            'sum': the output will be summed
+
+    Examples:
+        .. code-block:: python
+        import numpy as np
+        import paddle
+        from paddle import fluid
+        import paddle.fluid.dygraph as dg
+
+        mse_loss = paddle.nn.loss.MSELoss()
+        input = fluid.data(name="input", shape=[1])
+        label = fluid.data(name="label", shape=[1])
+        place = fluid.CPUPlace()
+        input_data = np.array([1.5]).astype("float32")
+        label_data = np.array([1.7]).astype("float32")
+
+        # declarative mode
+        output = mse_loss(input,label)
+        exe = fluid.Executor(place)
+        exe.run(fluid.default_startup_program())
+        output_data = exe.run(
+            fluid.default_main_program(),
+            feed={"input":input_data, "label":label_data},
+            fetch_list=[output],
+            return_numpy=True)
+        print(output_data)
+        # [array([0.04000002], dtype=float32)]
+
+        # imperative mode
+        with dg.guard(place) as g:
+            input = dg.to_variable(input_data)
+            label = dg.to_variable(label_data)
+            output = mse_loss(input, label)
+            print(output.numpy())
+            # [0.04000002]
+    """
+
+    def __init__(self, reduction='mean'):
+        super(MSELoss, self).__init__()
+        if reduction not in ['sum', 'mean', 'none']:
+            raise ValueError(
+                "'reduction' in 'MSELoss' should be 'sum', 'mean' or 'none', "
+                "but received {}.".format(reduction))
+        self.reduction = reduction
+
+    def forward(self, input, label):
+        if not fluid.framework.in_dygraph_mode():
+            fluid.data_feeder.check_variable_and_dtype(input, 'input',
+                                                       ['float32'], 'MSELoss')
+            fluid.data_feeder.check_variable_and_dtype(label, 'label',
+                                                       ['float32'], 'MSELoss')
+
+        helper = fluid.layer_helper.LayerHelper('MSELoss')
+        minus_out = helper.create_variable_for_type_inference(input.dtype)
+        helper.append_op(
+            type='elementwise_sub',
+            inputs={'X': [input],
+                    'Y': [label]},
+            outputs={'Out': [minus_out]})
+
+        square_out = helper.create_variable_for_type_inference(input.dtype)
+        helper.append_op(
+            type='square',
+            inputs={'X': [minus_out]},
+            outputs={'Out': [square_out]})
+
+        if self.reduction == 'none':
+            return square_out
+
+        reduce_op = 'reduce_mean'
+        if self.reduction == 'sum':
+            reduce_op = 'reduce_sum'
+
+        attrs = {'dim': [0], 'keep_dim': False, 'reduce_all': True}
+
+        if fluid.framework.in_dygraph_mode():
+            inputs = {'X': [square_out]}
+            outs = getattr(fluid.core.ops, reduce_op)(inputs, attrs)
+            return outs['Out'][0]
+
+        out = helper.create_variable_for_type_inference(input.dtype)
+        helper.append_op(
+            type=reduce_op,
+            inputs={'X': square_out},
+            outputs={'Out': out},
+            attrs=attrs)
+        return out
 
 
 class L1Loss(fluid.dygraph.Layer):
